+++ example/lib/routes/popups.dart
@@ -1,3 +1,4 @@
 export '../screens/popups/content_dialog.dart';
 export '../screens/popups/flyout.dart';
 export '../screens/popups/tooltip.dart';
+export '../screens/popups/snackbar.dart';