+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Next]
 
+- Add example page of `Snackbar` in /popups/snackbar
+
 - Add parameters `onTapDown` and `onTapUp` on all buttons. ([#795](https://github.com/bdlukaa/fluent_ui/pull/795))
    - **Breaking: if you use the abstract class `BaseButton`, these two parameters are now required** 
 - Add `PasswordBox` widget. ([#795](https://github.com/bdlukaa/fluent_ui/pull/795))